e2e2/test/images/neuron-training/requirements.txt
bokeh>=2.3,<3
awscli<2
scipy
click
cryptography
psutil==5.6.7
dataset
transformers==4.36.2
Pillow
torch-neuronx==2.1.2.2.3.0
neuronx-cc==2.15.128.0
neuronx_distributed==0.8.0
attrs<24,>=23.1.0
protobuf>=3.18.3,<=3.20.3
docutils>=0.10,<0.17
rsa<4.8,>=3.1.2
urllib3>=1.26.0,<1.27
torchvision==0.16.*
